refactor: :truck: atoms moved to data folder
