+++ src/components/PdfViewer.tsx
@@ -2,9 +2,9 @@ import { useAtom } from 'jotai';
 import { FC, useCallback, useEffect, useRef, useState } from 'react';
 import { VariableSizeList } from 'react-window';
 import useResizeObserver from 'use-resize-observer';
+import { scaleAtom } from '../data/atoms';
 import Page from './Page';
 import PdfPage from './PdfPage';
-import { scaleAtom } from './atoms';
 
 type PdfViewerProps = {
   width?: number | string;
